Added enemy health bar and added damage dropoff to push attack

+++ Assets/Scripts/Player/PushProjectile.cs
@@ -10,7 +10,7 @@ public class PushProjectile : MonoBehaviour
         Debug.Log(other);
         if (other.CompareTag("Enemy")) 
         {
-            other.GetComponent<DestroyOnHealth>().Health -= damage;
+            other.GetComponent<DestroyOnHealth>().Health -= (int)damage;
         }
     }
 }
